brain/requirements.txt
Django[argon2]>=5.2.0,<5.3.0
psycopg
django-environ
Pillow
pgvector

django-polymorphic


django-allauth>=0.60.0
django-countries[pyuca]
django-import-export[all]
django-storages[google]

django-imagekit
pilkit>=3.0.0

django-compressor>=4.5
beautifulsoup4
lxml

django-crispy-forms
crispy-bootstrap5
django-widget-tweaks

django-admin-rangefilter
django-filter

django-el-pagination

markdown
django-markdownify

requests

pycountry
pytimeparse2

celery
flower
django-celery-results
django-celery-beat

django-mailbox
python-social-auth
social-auth-app-django>=5.4.1

djangorestframework
markdown
django-filter
django-oauth-toolkit
drf-spectacular[sidecar]
drf-extra-fields

# Waiting for semanticscholar > 0.10.0 to be released
# https://github.com/danielnsilva/semanticscholar/issues/111
semanticscholar @ git+https://github.com/danielnsilva/semanticscholar/@master

arxiv

django-debug-toolbar>=5.2.0
